Replace scan-build static analysis with gated clang-tidy
The static target ran scan-build, which was broken: the image ships the
analyzer as scan-build-21 and the pinned-compiler preset defeats its
compiler interception. Switch to clang-tidy -- the same static-analysis
engine plus bugprone, modernize, and performance checks -- run directly
off compile_commands.json. The static preset disables precompiled headers
so the analyzer needs no build and lints header bodies the precompiled
header would otherwise hide.

Five checks gate the build: branch-clone, implicit widening of a
multiplication result, missing override, inefficient vector operation, and
mismatched argument comments. Clearing them touched 45 files: identical
switch branches merged, or annotated where the duplication is a deliberate
ISA or protocol table; implicit integer-multiply widenings made explicit;
missing override added; vectors reserved before push loops; and stale
argument-name comments corrected in tests. Two checks stay advisory because
they conflict with legitimate conventions across the tree: avoid C arrays
(valid in real-time and embedded code) and identifier naming (the tree
mixes math-symbol, prefixed-global, and padding-member naming that flat
case rules cannot express).

The nightly static job now runs clang-tidy. All source changes preserve
behavior.

Verified: make static reports zero gated findings; make debug and the full
suite pass 4405 of 4405 tests.

+++ src/system/core/components/scheduler/posix/ptst/SchedulerMultiThread_pTest.cpp
@@ -28,6 +28,7 @@
 #include <gtest/gtest.h>
 
 #include <chrono>
+#include <cstddef>
 #include <cstdint>
 #include <cstdio>
 #include <cstdlib>
@@ -252,9 +253,9 @@ PERF_TEST(SchedulerMtPerf, ManyNoWork) {
   // Enhanced memory profile accounting
   const size_t TASK_PTR = sizeof(SchedulableTask*);
   const size_t TASK_META = 64; // Cache line per task (metadata)
-  const size_t WORKER_STACK =
-      std::thread::hardware_concurrency() * 8192; // Conservative stack estimate
-  const size_t COHERENCY = N * 64;                // Cache line bouncing (pessimistic)
+  const size_t WORKER_STACK = static_cast<const size_t>(std::thread::hardware_concurrency() *
+                                                        8192); // Conservative stack estimate
+  const size_t COHERENCY = N * 64;                             // Cache line bouncing (pessimistic)
 
   ub::MemoryProfile memProfile{.bytesRead = N * (TASK_PTR + TASK_META) + WORKER_STACK + COHERENCY,
                                .bytesWritten = N * sizeof(std::uint8_t), // Result codes
@@ -323,7 +324,7 @@ PERF_TEST(SchedulerMtPerf, ManyLightWork) {
 
   // Memory profile with work included
   const size_t TASK_META = sizeof(SchedulableTask*) + 64;
-  const size_t WORKER_STACK = std::thread::hardware_concurrency() * 8192;
+  const size_t WORKER_STACK = static_cast<const size_t>(std::thread::hardware_concurrency() * 8192);
   const size_t COHERENCY = N * 64;
 
   ub::MemoryProfile memProfile{.bytesRead = N * TASK_META + WORKER_STACK + COHERENCY,
